+++ framework/domain/src/test/java/is/codion/framework/domain/TestDomain.java
@@ -118,11 +118,9 @@ public final class TestDomain extends DefaultDomain {
   void master() {
     add(Master.TYPE.define(
             Master.ID.define()
-                    .primaryKey()
-                    .beanProperty("id"),
+                    .primaryKey(),
             Master.NAME.define()
-                    .column()
-                    .beanProperty("name"),
+                    .column(),
             Master.CODE.define()
                     .column(),
             Master.READ_ONLY.define()
@@ -286,8 +284,7 @@ public final class TestDomain extends DefaultDomain {
   void detail() {
     add(Detail.TYPE.define(
             Detail.ID.define()
-                    .primaryKey()
-                    .beanProperty("id"),
+                    .primaryKey(),
             Detail.SHORT.define()
                     .column()
                     .caption(Detail.SHORT.name()),
@@ -297,8 +294,7 @@ public final class TestDomain extends DefaultDomain {
             Detail.DOUBLE.define()
                     .column()
                     .caption(Detail.DOUBLE.name())
-                    .columnHasDefaultValue(true)
-                    .beanProperty("double"),
+                    .columnHasDefaultValue(true),
             Detail.STRING.define()
                     .column()
                     .caption("Detail string")
@@ -325,14 +321,12 @@ public final class TestDomain extends DefaultDomain {
                     .column(),
             Detail.MASTER_FK.define()
                     .foreignKey()
-                    .caption(Detail.MASTER_FK.name())
-                    .beanProperty("master"),
+                    .caption(Detail.MASTER_FK.name()),
             Detail.MASTER_CODE_NON_DENORM.define()
                     .column(),
             Detail.MASTER_VIA_CODE_FK.define()
                     .foreignKey()
-                    .caption(Detail.MASTER_FK.name())
-                    .beanProperty("master"),
+                    .caption(Detail.MASTER_FK.name()),
             Detail.MASTER_NAME.define()
                     .denormalized(Detail.MASTER_FK, Master.NAME)
                     .caption(Detail.MASTER_NAME.name()),
@@ -383,24 +377,20 @@ public final class TestDomain extends DefaultDomain {
             Department.ID.define()
                     .primaryKey()
                     .caption(Department.ID.name())
-                    .updatable(true).nullable(false)
-                    .beanProperty("deptNo"),
+                    .updatable(true).nullable(false),
             Department.NAME.define()
                     .column()
                     .caption(Department.NAME.name())
                     .searchable(true)
                     .maximumLength(14)
-                    .nullable(false)
-                    .beanProperty("name"),
+                    .nullable(false),
             Department.LOCATION.define()
                     .column()
                     .caption(Department.LOCATION.name())
-                    .maximumLength(13)
-                    .beanProperty("location"),
+                    .maximumLength(13),
             Department.ACTIVE.define()
                     .booleanColumn(Integer.class, 1, 0)
-                    .readOnly(true)
-                    .beanProperty("active"),
+                    .readOnly(true),
             Department.DATA.define()
                     .column()
                     .lazy(true),
@@ -443,48 +433,39 @@ public final class TestDomain extends DefaultDomain {
             Employee.ID.define()
                     .primaryKey()
                     .caption(Employee.ID.name())
-                    .name("empno")
-                    .beanProperty("id"),
+                    .name("empno"),
             Employee.NAME.define()
                     .column()
                     .caption(Employee.NAME.name())
                     .searchable(true)
                     .name("ename")
                     .maximumLength(10)
-                    .nullable(false)
-                    .beanProperty("name"),
+                    .nullable(false),
             Employee.DEPARTMENT_NO.define()
                     .column()
-                    .nullable(false)
-                    .beanProperty("deptno"),
+                    .nullable(false),
             Employee.DEPARTMENT_FK.define()
                     .foreignKey()
-                    .caption(Employee.DEPARTMENT_FK.name())
-                    .beanProperty("department"),
+                    .caption(Employee.DEPARTMENT_FK.name()),
             Employee.JOB.define()
                     .column()
                     .items(asList(item("ANALYST"), item("CLERK"),
                             item("MANAGER"), item("PRESIDENT"), item("SALESMAN")))
                     .caption(Employee.JOB.name())
-                    .searchable(true)
-                    .beanProperty("job"),
+                    .searchable(true),
             Employee.SALARY.define()
                     .column()
                     .caption(Employee.SALARY.name())
-                    .nullable(false).valueRange(1000, 10000).maximumFractionDigits(2)
-                    .beanProperty("salary"),
+                    .nullable(false).valueRange(1000, 10000).maximumFractionDigits(2),
             Employee.COMMISSION.define()
                     .column()
                     .caption(Employee.COMMISSION.name())
-                    .valueRange(100, 2000).maximumFractionDigits(2)
-                    .beanProperty("commission"),
+                    .valueRange(100, 2000).maximumFractionDigits(2),
             Employee.MGR.define()
-                    .column()
-                    .beanProperty("mgr"),
+                    .column(),
             Employee.MANAGER_FK.define()
                     .foreignKey()
-                    .caption(Employee.MANAGER_FK.name())
-                    .beanProperty("manager"),
+                    .caption(Employee.MANAGER_FK.name()),
             Employee.HIREDATE.define()
                     .column()
                     .caption(Employee.HIREDATE.name())
@@ -493,8 +474,7 @@ public final class TestDomain extends DefaultDomain {
                             .delimiterDot()
                             .yearFourDigits()
                             .build())
-                    .nullable(false)
-                    .beanProperty("hiredate"),
+                    .nullable(false),
             Employee.DEPARTMENT_LOCATION.define()
                     .denormalized(Employee.DEPARTMENT_FK, Department.LOCATION)
                     .caption(Department.LOCATION.name()),